Remove GNU Emacs configuration
After trying out GNU Emacs again, I have come to the conclusion that it
is not for me. It doesn't seem worth it to switch from Vim, as I already
know so much, and have an almost static configuration at this point.

Emacs seems to encourage a culture of complexity, everything becomes
part of the editor, the packages are huge beasts which can do everything
and I would never use 90% of the features and key bindings. This became
apparent while looking at Clojure tooling.

For now I will stay with Vim, if I ever switch editors it will likely be
one which I have created (I already have some ideas for one).